-key option and pure exporting feature
